requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
cffi==1.14.0
chardet==3.0.4
idna==2.10
multidict==4.7.6
pycparser==2.20
PyNaCl==1.4.0
python-dotenv==0.14.0
six==1.15.0
sounddevice==0.3.15
websockets==8.1
yarl==1.4.2
